endpoint CRUD cho quản lý phiếu

+++ user_service/user_service/src/main/java/com/example/user_service/repository/RoleRepository.java
@@ -4,6 +4,6 @@ import com.example.user_service.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> { // sửa Long
     Optional<Role> findByRoleName(String roleName);
 }
